clients/simple-vulkan: New Vulkan client example
Example to serve as a reference of a basic pure Wayland client using
Vulkan, contained in a single source file and with a minimal set of
dependencies.
It heavily borrows from weston-simple-egl being a reference for EGL
and GLES Wayland clients. The source code aims to stay similar to
simple-egl.c.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>